requirements.txt
pycairo>=1.23.0
svg.py>=1.4.2


